+++ lib/target/target.dart
@@ -1,3 +1,5 @@
+import 'dart:collection';
+
 import '../excel/excel.dart';
 import '../exporter/exporter.dart';
 import '../settings.dart';
@@ -7,6 +9,8 @@ import 'localizations.dart';
 
 abstract class Target {
   factory Target(L10nSheet sheetData) {
+    Target? result;
+
     final targetNode = Settings.map['target'];
     final name = switch (targetNode) {
       final String e => e,
@@ -14,29 +18,55 @@ abstract class Target {
       _ => throw const _TargetSettingsError(),
     };
 
+    T possibleTargetValue<T>(String key, {required T defaultValue}) {
+      if (targetNode is MapBase) {
+        return targetNode[key] ?? defaultValue;
+      } else {
+        return defaultValue;
+      }
+    }
+
     try {
       switch (name) {
         case 'getx':
-          return GetX.withSheetData(sheetData);
+          result = GetX.withSheetData(sheetData);
         case 'arb':
-          return Arb.withSheetData(
+          result = Arb.withSheetData(
             sheetData,
-            genL10nYaml: switch (targetNode) {
-                  {'genL10nYaml': final bool? e} => e,
-                  _ => null,
-                } ??
-                false,
+            genL10nYaml: possibleTargetValue(
+              'genL10nYaml',
+              defaultValue: false,
+            ),
           );
         case 'localizations':
-          return Localizations.withSheetData(sheetData);
+          result = Localizations.withSheetData(
+            sheetData,
+            className: possibleTargetValue(
+              'className',
+              defaultValue: 'L',
+            ),
+            outputFileName: possibleTargetValue(
+              'outputFileName',
+              defaultValue: 'app_localizations',
+            ),
+            genExtension: possibleTargetValue(
+              'genExtension',
+              defaultValue: false,
+            ),
+          );
       }
     } catch (e) {
       throw _TargetSettingsError(name);
     }
-
-    throw UnimplementedError('Unsupported target: $name');
+    if (result == null) {
+      throw UnimplementedError('Unsupported target: $name');
+    }
+    current = result;
+    return result;
   }
   const Target.withSheetData(this.sheetData);
+
+  static late final Target current;
 
   final L10nSheet sheetData;
   List<BaseExporter> get exporters;